requirements.txt
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
